fix(lunas.dataset): fix bugs in loading and saving datasets inherited from NestedN and bug in Chunk
To enable checkpointing, the descendant datasets run a complete iteration and then the root dataset
uses ptr to skip elements to resume from previous checkpointed state. This can incur significant
overhead to resume the iteration if previously stopped near the end of a large dataset or the
subsequent datasets carry heavy mapping function. `Window` is removed since it seems pointless
currently. An error in `Chunk` is fixed.